AC: support segnet model
