+++ tools/accuracy_checker/accuracy_checker/annotation_converters/__init__.py
@@ -105,7 +105,7 @@ from .common_object_detection import CommonDetectionConverter
 from .wflw import WFLWConverter
 from .see_in_the_dark import SeeInTheDarkDatasetConverter
 from .conll_ner import CONLLDatasetConverter
-from .background_matting import BackgroundMattingConverter
+from .background_matting import BackgroundMattingConverter, VideoBackgroundMatting
 from .tacotron2_test_data_converter import TacotronDataConverter
 from .noise_suppression_dataset import NoiseSuppressionDatasetConverter
 from .vimeo90k_sr import Vimeo90KSuperResolutionDatasetConverter
@@ -218,6 +218,7 @@ __all__ = [
     'SeeInTheDarkDatasetConverter',
     'CONLLDatasetConverter',
     'BackgroundMattingConverter',
+    'VideoBackgroundMatting',
     'TacotronDataConverter',
     'NoiseSuppressionDatasetConverter',
     'Vimeo90KSuperResolutionDatasetConverter',
