+++ tools/accuracy_checker/accuracy_checker/data_readers/__init__.py
@@ -18,26 +18,6 @@ from .data_reader import (
     BaseReader,
     DataReaderField,
     ReaderCombiner,
-    OpenCVFrameReader,
-    OpenCVImageReader,
-    PillowImageReader,
-    ScipyImageReader,
-    NiftiImageReader,
-    NumPyReader,
-    NumpyTXTReader,
-    NumpyDictReader,
-    NumpyBinReader,
-    TensorflowImageReader,
-    AnnotationFeaturesReader,
-    WavReader,
-    DicomReader,
-    PickleReader,
-    SkimageReader,
-    RawpyReader,
-    ByteFileReader,
-    LMDBReader,
-    KaldiARKReader,
-
     DataRepresentation,
     ClipIdentifier,
     MultiFramesInputIdentifier,
@@ -47,6 +27,7 @@ from .data_reader import (
     KaldiFrameIdentifier,
     KaldiMatrixIdentifier,
     ParametricImageIdentifier,
+    VideoFrameIdentifier,
 
     serialize_identifier,
     deserialize_identifier,
@@ -55,12 +36,37 @@ from .data_reader import (
     create_reader,
     REQUIRES_ANNOTATIONS
 )
+from .annotation_readers import AnnotationFeaturesReader
+from .binary_data_readers import PickleReader, ByteFileReader, LMDBReader
+from .medical_imaging_readers import NiftiImageReader, DicomReader
+from .audio_readers import WavReader, KaldiARKReader
+from .numpy_readers import NumPyReader, NumpyTXTReader, NumpyDictReader, NumpyBinReader
+from .image_readers import (
+    OpenCVImageReader,
+    PillowImageReader,
+    ScipyImageReader,
+    OpenCVFrameReader,
+    TensorflowImageReader,
+    SkimageReader,
+    RawpyReader
+)
 
 __all__ = [
     'BaseReader',
     'DataReaderField',
     'DataRepresentation',
     'ReaderCombiner',
+    'DataRepresentation',
+    'ClipIdentifier',
+    'MultiFramesInputIdentifier',
+    'ImagePairIdentifier',
+    'ListIdentifier',
+    'MultiInstanceIdentifier',
+    'KaldiMatrixIdentifier',
+    'KaldiFrameIdentifier',
+    'ParametricImageIdentifier',
+    'VideoFrameIdentifier',
+
     'OpenCVFrameReader',
     'OpenCVImageReader',
     'PillowImageReader',
@@ -80,16 +86,6 @@ __all__ = [
     'ByteFileReader',
     'LMDBReader',
     'KaldiARKReader',
-
-    'DataRepresentation',
-    'ClipIdentifier',
-    'MultiFramesInputIdentifier',
-    'ImagePairIdentifier',
-    'ListIdentifier',
-    'MultiInstanceIdentifier',
-    'KaldiMatrixIdentifier',
-    'KaldiFrameIdentifier',
-    'ParametricImageIdentifier',
 
     'create_reader',
     'REQUIRES_ANNOTATIONS',
